+++ CHANGELOG.rst
@@ -12,6 +12,8 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 * Annotate most ``AssertionManager`` methods using Protocols.
 * Moved Protocols to their own separate stub module.
 * Cleaned up the ``_MetaAM`` metaclass.
+* Reworked some of the internals of ``AssertionManager``.
+* Added tests using `pydocstyle <https://github.com/henry0312/pytest-pydocstyle>`_.
 
 
 2.2.3